+++ R/targets.R
@@ -39,9 +39,14 @@ use_targets <- function() {
 #' @param params A [data.frame] or [list] contains the parameters to be bound to
 #'   the query. Default templates require specifying `organization_name` and
 #'   `project_name`, in that order. If `contents` template is specified without
-#'   any parameters, set it as empty vector or `NULL`. If `contents` template is
-#'   specified as a [data.frame] as is, this parameter is omitted.
+#'   any parameters, set it as empty vector or `NULL`. If `contents` argument is
+#'   specified, this argument is omitted.
 #' @param ... For future usage. Should be empty.
+#' @param contents The contents used as the configuration of data fetching. This
+#'   is typically automatically fetched from database based on the `contents`
+#'   template from `templates`. If not `NULL`, it will be used directly and
+#'   ignore that specified in `templates`. Note `contents` should at least
+#'   contains `project_id`, `game_id` and `course_date` names.
 #' @param what What to fetch. If set as "all", both raw data and scores will be
 #'   fetched. If set as "raw_data", only raw data will be fetched. If set as
 #'   "scores", only scores will be fetched. Further actions on the fetched raw
@@ -57,6 +62,7 @@ use_targets <- function() {
 #' @return A list of target objects.
 #' @export
 prepare_fetch_data <- function(params, ...,
+                               contents = NULL,
                                what = c("all", "raw_data", "scores"),
                                action_raw_data = c("all", "parse", "none"),
                                templates = setup_templates(),
@@ -71,9 +77,7 @@ prepare_fetch_data <- function(params, ...,
   what <- match.arg(what)
   if (what == "all") what <- c("raw_data", "scores")
   action_raw_data <- match.arg(action_raw_data)
-  if (inherits(templates$contents, "AsIs")) {
-    contents <- templates$contents
-  } else {
+  if (is.null(contents)) {
     contents <- fetch_iquizoo_mem(
       read_file(templates$contents),
       params = unname(
@@ -118,8 +122,7 @@ prepare_fetch_data <- function(params, ...,
 #'   `project_id`, `game_id` and `course_date` should be included in the fetched
 #'   data based on the template. `project_id` will be used as the only parameter
 #'   in `users` and `project` templates, while all three will be used in
-#'   `raw_data` and `scores` templates. Call [I()] to specify a [data.frame] if
-#'   you have pre-fetched contents.
+#'   `raw_data` and `scores` templates.
 #' @param users The SQL template file used to fetch users. Usually you don't
 #'   need to change this.
 #' @param raw_data The SQL template file used to fetch raw data. See